Use float instead of double

+++ src/components/physics.h
@@ -3,7 +3,7 @@
 #include "vec2/vec2.h"
 
 struct Physics {
-  double speed = 0.0;
+  float speed = 0.0;
   Vec2 velocity = Vec2();
   Vec2 acceleration = Vec2();
 };
