+++ src/vec2/vec2.h
@@ -2,23 +2,23 @@
 
 class Vec2 {
  public:
-  double x, y;
+  float x, y;
 
   Vec2();
 
-  Vec2(double x, double y);
+  Vec2(float x, float y);
 
   static Vec2 add(Vec2 u, Vec2 v);
 
   static Vec2 subtract(Vec2 u, Vec2 v);
 
-  static Vec2 multiply(Vec2 v, double n);
+  static Vec2 multiply(Vec2 v, float n);
 
-  static Vec2 divide(Vec2 v, double n);
+  static Vec2 divide(Vec2 v, float n);
 
-  static double magnitude(Vec2 v);
+  static float magnitude(Vec2 v);
 
   static Vec2 normalize(Vec2 v);
 
-  static Vec2 limit(Vec2 v, double n);
+  static Vec2 limit(Vec2 v, float n);
 };
